Admin files
New backend role: the ‘File’ role. What does that mean? Quite simply: a fully-fledged file server with authorisation controls. Administrators now have the option to securely make tools and files available for download for internal or external purposes.
But that's not all - the ‘File’ role can also be used, for example, for automated file downloads directly via PowerShell on client systems, as shown in the following example.

Step 1: Create a directory, upload the Procmon.exe and copy the link.
Step 2: Using the remote shell of the test device, we download the Procmon.exe directly from our file server to the target device. This makes remote scripting even more efficient.

This makes managing tools and files easier and brings more fun to the workflow!